Translate to jUnit test cases
